// packages/common/definitions.d.ts
export interface Dump {
    artefact_of_power_group_keywords_keyword: ArtefactOfPowerGroupKeywordsKeyword[];
    conditional_role_general_keywords_keyword: ConditionalRoleLKeywordsKeyword[];
    battalion_ability: BattalionAbility[];
    conditional_role: ConditionalRole[];
    spell_lore_group_required_keywords_keyword: SpellLoreGroupRequiredKeywordsKeyword[];
    mount_trait_keywords_keyword: MountTraitKeywordsKeyword[];
    coalition_ally_option: CoalitionAllyOption[];
    core_battalion_group: CoreBattalionGroupElement[];
    core_battalion: CoreBattalion[];
    battle_trait_excludes_sub_factions_faction: BattleTraitExcludesSubFactionsFaction[];
    command_trait_group_keywords_keyword: CommandTraitGroupKeywordsKeyword[];
    conditional_role_required_warscroll_keywords_keyword: ConditionalRoleLKeywordsKeyword[];
    conditional_role_battlefield_role: ConditionalRoleBattlefieldRole[];
    artefact_of_power_group: PurpleGroup[];
    battle_trait_subtrait: {
        [key: string]: null | string;
    }[];
    artefact_of_power: ArtefactOfPower[];
    command_trait_group: PurpleGroup[];
    enhancement_bonus_requires_keywords_keyword: EnhancementBonusEsKeywordsKeyword[];
    battle_trait_group: PurpleGroup[];
    grand_strategy: GrandStrategy[];
    faction_allies_faction: FactionAlliesFaction[];
    prayer_group: PurpleGroup[];
    faction: Faction[];
    mount_trait_group: PurpleGroup[];
    migrations: Migration[];
    mount_trait_group_keywords_keyword: MountTraitGroupKeywordsKeyword[];
    grand_strategy_army_must_contain_keywords_keyword: GrandStrategyArmyMustContainKeywordsKeyword[];
    grand_strategy_keywords_keyword: {}[];
    mount_trait: {
        [key: string]: null | string;
    }[];
    grand_strategy_group: CoreBattalionGroupElement[];
    keyword: Keyword[];
    enhancement_bonus: EnhancementBonus[];
    grand_strategy_general_must_have_factions_faction: GrandStrategyGeneralMustHaveFactionsFaction[];
    command_trait: {
        [key: string]: null | string;
    }[];
    enhancement_bonus_excludes_keywords_keyword: EnhancementBonusEsKeywordsKeyword[];
    battle_trait: BattleTrait[];
    role_modifier: RoleModifier[];
    triumph_group: PurpleGroup[];
    unique_enhancement_group: CoreBattalionGroupElement[];
    unit_type_excluded_battlefield_role: UnitTypeEdBattlefieldRole[];
    prayer: {
        [key: string]: null | string;
    }[];
    unit_type_required_battlefield_role: UnitTypeEdBattlefieldRole[];
    unique_enhancement_group_keywords_keyword: UniqueEnhancementGroupKeywordsKeyword[];
    warscroll: Warscroll[];
    damage_row: DamageRow[];
    unit_type_required_keywords_keyword: UnitTypeRequiredKeywordsKeyword[];
    unique_enhancement_keywords_keyword: UniqueEnhancementKeywordsKeyword[];
    unique_enhancement: {
        [key: string]: null | string;
    }[];
    version_info: VersionInfo[];
    unit_type: UnitType[];
    publication: Publication[];
    spell_lore: {
        [key: string]: null | string;
    }[];
    spell_lore_group: PurpleGroup[];
    triumph: {
        [key: string]: null | string;
    }[];
    description_subsection: DescriptionSubsection[];
    warscroll_keywords_keyword: WarscrollKeywordsKeyword[];
    damage_cell: DamageCell[];
    warscroll_excludes_other_warscrolls_warscroll: WarscrollExcludesOtherWarscrollsWarscroll[];
    warscroll_requires_other_warscrolls_warscroll: WarscrollRequiresOtherWarscrollsWarscroll[];
    warscroll_mark_of_chaos_options_faction: WarscrollTionsFaction[];
    warscroll_battlefield_role: WarscrollBattlefieldRole[];
    weapon: Weapon[];
    warscroll_factions_faction: WarscrollTionsFaction[];
    warscroll_ability: WarscrollAbility[];
}
export interface WarscrollAbility {
    id: string;
    name: string;
    lore: string;
    rules: string;
    warscrollId: string;
}
export interface ArtefactOfPower {
    id: string;
    name: string;
    lore: string;
    rules: string;
    requiresMeleeWeapon: boolean;
    requiresMissileWeapon: boolean;
    givesWizard: boolean;
    artefactOfPowerGroupId: string;
}
export interface PurpleGroup {
    id: string;
    name: string;
    restrictions?: string;
    publicationId: string;
    factionId: null | string;
    allowUniqueUnits?: boolean;
}
export interface ArtefactOfPowerGroupKeywordsKeyword {
    artefactOfPowerGroupId: string;
    keywordId: string;
}
export interface BattalionAbility {
    id: string;
    header: string;
    rules: string;
    grantsExtraEnhancement: boolean;
    coreBattalionId: string;
}
export interface BattleTrait {
    id: string;
    name: string;
    lore: string;
    rules: string;
    additionalArtefacts: number;
    battleTraitGroupId: string;
}
export interface BattleTraitExcludesSubFactionsFaction {
    battleTraitId: string;
    factionId: string;
}
export interface CoalitionAllyOption {
    id: string;
    xInFourLimit: number;
    factionId: string;
    requiredFactionId: string;
    excludedFactionId: null;
    requiredMarkOfChaosId: null;
}
export interface CommandTraitGroupKeywordsKeyword {
    commandTraitGroupId: string;
    keywordId: string;
}
export interface ConditionalRole {
    id: string;
    enforced: boolean;
    rosterFactionId: null | string;
    rosterTypeId: null | string;
    rosterSubfactionId: null | string;
    generalFactionId: null | string;
    allUnitsFactionId: null | string;
    warscrollId: string;
}
export interface ConditionalRoleBattlefieldRole {
    role: Role;
    conditionalRoleId: string;
}
export declare enum Role {
    Artillery = "artillery",
    Battleline = "battleline",
    Behemoth = "behemoth",
    EndlessSpell = "endlessSpell",
    Invocation = "invocation",
    Leader = "leader",
    Other = "other",
    Terrain = "terrain"
}
export interface ConditionalRoleLKeywordsKeyword {
    conditionalRoleId: string;
    keywordId: string;
}
export interface CoreBattalion {
    id: string;
    name: string;
    onePerArmy: boolean;
    coreBattalionGroupId: string;
}
export interface CoreBattalionGroupElement {
    id: string;
    name: string;
    rules: string;
    requiresBattlepack?: null | string;
    factionId: null | string;
    publicationId: string;
    restrictions?: string;
    domain?: string;
}
export interface DamageCell {
    id: string;
    order: number;
    text: string;
    rowId: string;
}
export interface DamageRow {
    id: string;
    order: number;
    warscrollId: string;
}
export interface DescriptionSubsection {
    id: string;
    header: string;
    rules: string;
    warscrollId: string;
}
export interface EnhancementBonus {
    id: string;
    bonusType: string;
    modifier: number;
    overridesLimit: boolean;
    requiresBearerItem: boolean;
    scope: Scope;
    artefactId: null | string;
    commandTraitId: null | string;
    battleTraitId: null | string;
    battleTraitSubtraitId: null | string;
    bonusArtefactGroupId: null | string;
    bonusSpellGroupId: null | string;
    bonusPrayerGroupId: null | string;
    bonusUniqueEnhancementGroupId: null | string;
    gainsKeywordId: null | string;
    bonusBattleTraitId: null | string;
}
export declare enum Scope {
    Army = "army",
    Bearer = "bearer",
    Unit = "unit"
}
export interface EnhancementBonusEsKeywordsKeyword {
    enhancementBonusId: string;
    keywordId: string;
}
export interface Faction {
    id: string;
    name: string;
    keywordCategory: KeywordCategory;
    restrictive: boolean;
    parentId: null | string;
    requiredGeneralKeywordId: null | string;
    confersToCoalitionAllies: boolean;
}
export declare enum KeywordCategory {
    ArmyType = "armyType",
    Generic = "generic",
    RosterLevel = "rosterLevel",
    Subfaction = "subfaction"
}
export interface FactionAlliesFaction {
    factionId: string;
    alliedFactionId: string;
}
export interface GrandStrategy {
    id: string;
    name: string;
    rules: string;
    grandStrategyGroupId: string;
    factionId: null | string;
}
export interface GrandStrategyArmyMustContainKeywordsKeyword {
    grandStrategyId: string;
    keywordId: string;
}
export interface GrandStrategyGeneralMustHaveFactionsFaction {
    grandStrategyId: string;
    factionId: string;
}
export interface Keyword {
    id: string;
    name: string;
}
export interface Migration {
    id: number;
    timestamp: string;
    name: string;
}
export interface MountTraitGroupKeywordsKeyword {
    mountTraitGroupId: string;
    keywordId: string;
}
export interface MountTraitKeywordsKeyword {
    mountTraitId: string;
    keywordId: string;
}
export interface Publication {
    id: string;
    name: string;
    edition: number;
    productId: null | string;
}
export interface RoleModifier {
    id: string;
    role: Role;
    limitModifier: number;
    requiredAllUnitsKeywordId: null | string;
    battleTraitId: string;
}
export interface SpellLoreGroupRequiredKeywordsKeyword {
    spellLoreGroupId: string;
    keywordId: string;
}
export interface UniqueEnhancementGroupKeywordsKeyword {
    uniqueEnhancementGroupId: string;
    keywordId: string;
}
export interface UniqueEnhancementKeywordsKeyword {
    uniqueEnhancementId: string;
    keywordId: string;
}
export interface UnitType {
    id: string;
    name: string;
    restrictions: string;
    imageName: string;
    min: number;
    max: number;
    eitherOr: boolean;
    woundsLimit: number | null;
    coreBattalionId: string;
    order: number;
}
export interface UnitTypeEdBattlefieldRole {
    role: Role;
    unitTypeId: string;
}
export interface UnitTypeRequiredKeywordsKeyword {
    unitTypeId: string;
    keywordId: string;
}
export interface VersionInfo {
    id: string;
    dataVersion: number;
}
export interface Warscroll {
    id: string;
    name: string;
    subname: null | string;
    lore: string;
    points: number;
    unique: boolean;
    single: boolean;
    legends: boolean;
    unitSize: number | null;
    move: null | string;
    save: number | null;
    bravery: number | null;
    wounds: number | null;
    basicDescription: null | string;
    notes: null | string;
    referenceKeywords: string;
    ineligibleForGeneral: boolean;
    imageUrl: null | string;
    publicationId: string;
    markOfChaosRequired: boolean;
}
export interface WarscrollBattlefieldRole {
    role: Role;
    warscrollId: string;
}
export interface WarscrollExcludesOtherWarscrollsWarscroll {
    warscrollId: string;
    excludedWarscrollId: string;
}
export interface WarscrollTionsFaction {
    warscrollId: string;
    factionId: string;
}
export interface WarscrollKeywordsKeyword {
    warscrollId: string;
    keywordId: string;
}
export interface WarscrollRequiresOtherWarscrollsWarscroll {
    warscrollId: string;
    requiredWarscrollId: string;
}
export interface Weapon {
    id: string;
    name: string;
    type: Type;
    range: string;
    attacks: string;
    hit: string;
    wound: string;
    rend: string;
    damage: string;
    warscrollId: string;
}
export declare enum Type {
    Melee = "melee",
    Missile = "missile"
}